requirements.txt
# Core Data Processing
polars>=0.20.0
numpy>=1.24.0

# Web Framework (Streamlit)
streamlit>=1.28.0

# Data Visualization
plotly>=5.17.0

# HTTP Requests
requests>=2.31.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0

# Enhanced ML Models (optional but recommended)
xgboost>=2.0.0
lightgbm>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time Utilities (usually included with Python, but explicit for clarity)
# datetime - built-in
# typing - built-in

# System/OS utilities (built-in)
# os - built-in
# sys - built-in
# pathlib - built-in
# argparse - built-in
# logging - built-in
# traceback - built-in

# Optional: Additional useful packages for financial analysis
# Uncomment if you want to add more financial analysis capabilities
# yfinance>=0.2.0
# ta>=0.10.0
# quantlib>=1.32

# Development and Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0

# Performance optimization (optional)
# numba>=0.58.0 